+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- [89] Add retry mechanism when pulling image metadata to avoid installation/upgrade interrupts when the network is not
+available.
 
 ## [v0.27.0] - 2023-03-27
 ### Added